Move convert_* and fixed conversions to utils.rs
