+++ zrml/rikiddo/fuzz/fixedu_to_fixedi_conversion.rs
@@ -7,7 +7,7 @@
 
 use libfuzzer_sys::fuzz_target;
 use substrate_fixed::{types::extra::U33, FixedI128};
-use zrml_rikiddo::types::convert_to_signed;
+use zrml_rikiddo::utils::convert_to_signed;
 
 mod shared;
 use shared::fixed_from_u128;
